backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
sqlalchemy
passlib[bcrypt]
python-multipart
python-dotenv
google-genai
PyPDF2
Pillow
python-docx
fpdf2
